requirements.txt
psycopg2-binary>=2.9
mysqlclient>=2.2
passlib>=1.7
bcrypt>=4.0,<4.1
zeroc-ice>=3.7,<3.8.0
django>=4.2,<5.0
cryptography>=42.0
